+++ frontend/src/hooks/useScroll.js
@@ -1,0 +1,32 @@
+import React, { useState, useEffect } from "react";
+
+const useScrollEffect = () => {
+    const [showScrollButton, setShowScrollButton] = useState(false);
+
+    const scrollToTop = () => {
+        window.scrollTo({ top: 0, behavior: "smooth" });
+    };
+
+    const scrollToBottom = () => {
+        window.scrollTo({ top: document.body.offsetHeight, behavior: "smooth" });
+    };
+
+    useEffect(() => {
+        const handleScroll = () => {
+            if (window.scrollY > 100) {
+                setShowScrollButton(true);
+            } else {
+                setShowScrollButton(false);
+            }
+        };
+
+        window.addEventListener("scroll", handleScroll);
+        return () => {
+            window.removeEventListener("scroll", handleScroll);
+        };
+    }, []);
+
+    return { showScrollButton, scrollToTop, scrollToBottom };
+};
+
+export default useScrollEffect;
