+++ frontend/src/pages/Experience.jsx
@@ -1,4 +1,5 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
+
 import { VerticalTimeline, VerticalTimelineElement } from "react-vertical-timeline-component";
 import IconSchool from "@material-ui/icons/School";
 import IconWork from "@material-ui/icons/Work";
@@ -6,34 +7,15 @@ import IconInt from "@material-ui/icons/LocalLibraryOutlined";
 import { IoMdPulse } from "react-icons/io";
 import ArrowUpwardIcon from '@mui/icons-material/ArrowUpward';
 import ArrowDownwardIcon from '@mui/icons-material/ArrowDownward';
+
+import useScroll from "../hooks/useScroll";
+
 import "react-vertical-timeline-component/style.min.css";
 import "../styles/Experience.css";
 
+
 const Experience = () => {
-    const [showScrollButton, setShowScrollButton] = useState(false);
-
-    useEffect(() => {
-        const handleScroll = () => {
-            if (window.scrollY > 100) {
-                setShowScrollButton(true);
-            } else {
-                setShowScrollButton(false);
-            }
-        };
-
-        window.addEventListener("scroll", handleScroll);
-        return () => {
-            window.removeEventListener("scroll", handleScroll);
-        };
-    }, []);
-
-    const scrollToTop = () => {
-        window.scrollTo({ top: 0, behavior: "smooth" });
-    };
-
-    const scrollToBottom = () => {
-        window.scrollTo({ top: document.body.offsetHeight, behavior: "smooth" });
-    };
+    const { showScrollButton, scrollToTop, scrollToBottom } = useScroll();
 
     const calculateAge = (birthdate) => {
         const today = new Date();
