+++ frontend/src/pages/Skills.jsx
@@ -1,14 +1,20 @@
 import React, { useState, useEffect } from "react";
-import SkillsFilter from "../components/SkillsFilter";
-import SpinComponent from "../components/SpinComponent";
+
 import { Accordion, AccordionSummary, Typography, AccordionDetails } from "@mui/material";
 import ArrowUpwardIcon from '@mui/icons-material/ArrowUpward';
 import ArrowDownwardIcon from '@mui/icons-material/ArrowDownward';
 import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
+
+import SkillsFilter from "../components/SkillsFilter";
+import SpinComponent from "../components/SpinComponent";
+
+import useScroll from "../hooks/useScroll";
 import { getSkillList } from "../apis/SkillList";
 import "../styles/Home.css";
 
 const Skills = () => {
+    const { showScrollButton, scrollToTop, scrollToBottom } = useScroll();
+
     const [skills, setSkills] = useState([]);
     const [isLoading, setIsLoading] = useState(true);
     useEffect(() => {
@@ -19,30 +25,6 @@ const Skills = () => {
         };
         fetchData();
     }, []);
-
-    const [showScrollButton, setShowScrollButton] = useState(false);
-    useEffect(() => {
-        const handleScroll = () => {
-            if (window.scrollY > 100) {
-                setShowScrollButton(true);
-            } else {
-                setShowScrollButton(false);
-            }
-        };
-
-        window.addEventListener("scroll", handleScroll);
-        return () => {
-            window.removeEventListener("scroll", handleScroll);
-        };
-    }, []);
-
-    const scrollToTop = () => {
-        window.scrollTo({ top: 0, behavior: "smooth" });
-    };
-
-    const scrollToBottom = () => {
-        window.scrollTo({ top: document.body.offsetHeight, behavior: "smooth" });
-    };
 
     const accordionStyle = {
         borderRadius: "10px",
